+++ samples/server/petstore/jaxrs-cxf/src/test/java/io/swagger/api/StoreApiTest.java
@@ -113,7 +113,7 @@ public class StoreApiTest {
      */
     @Test
     public void getOrderByIdTest() {
-        Integer orderId = null;
+        Long orderId = null;
         //Order response = api.getOrderById(orderId);
         //assertNotNull(response);
         // TODO: test validations
